# REF/PERF: move np.errstate out of core array_ops up to higher level