@jbrockmendel 2021-03-12
This is unfortunate, since it is both a) uglier to do this at a higher level and b) a really nice perf improvement