@jorisvandenbossche 2021-03-12
Is it that much uglier? I removed it in 5 places, and I added it in 4 places in `_dispatch_frame_op` (which in theory could be combined into a single one) and 3 other places. So it's not that big of a difference.